+++ packages/icon/src/css/index.ts
@@ -6,7 +6,7 @@ import {
   colorsYellow
 } from '@pluralsight/ps-design-system-core'
 
-import * as vars from '../vars'
+import * as vars from '../vars/index.js'
 
 export default {
   [`.psds-icon`]: {
