+++ packages/icon/src/index.ts
@@ -1,4 +1,4 @@
-export { default as css } from './css'
-export * from './vars'
-export { default } from './react'
-export * from './react/icons'
+export { default as css } from './css/index.js'
+export * from './vars/index.js'
+export { default } from './react/index.js'
+export * from './react/icons/index.js'
